backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0

# OTP & Communication
twilio==8.11.1
python-dotenv==1.0.0

# Media
cloudinary==1.37.0

# Utilities
aiofiles==23.2.1
python-dateutil==2.8.2

# WebSockets
websockets==12.0

# Background Tasks
celery==5.3.6
celery[redis]==5.3.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.1.1
flake8==7.0.0
slowapi==0.1.9
bleach==6.1.0
